# Use gometalinter for linting