--- a/Makefile.windows
+++ b/Makefile.windows
@@ -1,7 +1,6 @@
 #Windows specific settings.
 WHALE = "+"
 ONI = "-"
-FIX_PATH = $(subst /,\,$1)
 
 BINARY_SUFFIX=".exe"