If you're interested, this PR uses gometalinter to run all the existing linters.

[gometalinter](https://github.com/alecthomas/gometalinter)  provides:
* exits with non-zero on failure (currently being handled by `test -z`)
* uniform mechanisms for ignoring/whitelisting files or issues (`// nolint` in the source, or the `Exclude` list in the config)
* runs all the linters in parallel
* installs a vendored version of all the linters with one command