@mlaventure 2017-11-03
You'll need to update `.appvveyor.yml` to add the go/bin path before calling on `make setup`