@dnephin 2017-11-03
Thanks, fixed the appveyor build. I'm seeing a similar error on the travis/darwin build as well, but it doesn't look like the previous travis config did anything with path.

linux is working, but failing on the next step with: `/bin/sh: 1: protobuild: not found`

I'm still looking into why