@dnephin 2017-11-03
From what I can tell, `lint` and `ineffassign` in the darwin build were failing silently because `GOOS=darwin` was set, which meant that `go get` was installing a cross-compiled version into `/go/bin/darwin_amd64`.

This is now fixed because `gometalinter` reports errors if the binary for a linter does not exist. It hit a few lint errors in some `_unix.go` and `bsd.go` files, which I'll fix to get this green.