@dqminh 2017-11-06
LGTM

```
master > time make check
🇩 fmt
🇩 vet
🇩 lint
🇩 ineffassign
make check  8.34s user 0.78s system 128% cpu 7.124 total

pr/1717 > time make check
🇩 proto-fmt
🇩 check
gometalinter --config .gometalinter.json ./...
make check  10.38s user 1.74s system 248% cpu 4.881 total
```

Its even faster !